pyproject.toml
[tool.poetry]
name = "common"
version = "0.1.0"
description = "A collection of utilities"
authors = ["Dev Tools <devtools@acme.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
opentelemetry-sdk = "^1.23.0"
opentelemetry-api = "^1.23.0"
setuptools = "^69.2.0"


[tool.poetry.group.dev.dependencies]
dumb-pypi = "^1.15.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
